requirements.txt
pandas==2.0.0
openpyxl==3.1.2
numpy==1.24.2
requests==2.29.0
pytz>=2022.1
python-dateutil~=2.8.2
sqlalchemy==2.0.14
yfinance==0.2.18
black=>23.1.0
pytest==7.3.1